fixed a dumb bug

+++ pastis.rb
@@ -36,8 +36,7 @@ class Pastis
   DEFAULT_FEED        = "http://www.ezrss.it/feed/" #"http://pipes.yahoo.com/pipes/pipe.run?_id=7aa6281616ea0a8cb27aaa0914f09a76&_render=rss"
   TORRENTS_LOCAL_PATH = File.expand_path("./torrents/")     # AKA the glass
   PRUNE_FILES_AFTER   = Time.now - (60 * 60 * 24 * 31)                                              
-  # TODO check the file exist
-  raise "You need to have a filters.yml file, check the example file"
+  raise "You need to have a filters.yml file, check the example file" unless File.exist?(ROOT + "/filters.yml")
   FILTERS             = YAML.load(File.open(ROOT + '/filters.yml', 'r')).map{|raw| ::Pastis::Filter.new(raw)}
    
   attr_reader :client, :server
